+++ changelog.md
@@ -5,13 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
  
+## Unreleased - 2023-03-15
+ 
+This change adds a script to gather the list of protobuf imports and then adds them into the readme.
+ 
+### Added
+- *arch3-proto*: added script to gather protobuf import [#39](https://github.com/archway-network/arch3.js)
+ 
+### Changed
+ 
+### Fixed
+
+
 ## Unreleased
  
 This change is a workspace addition of protobufs codegen capabilities. The new package arch3-proto is the core change.
  
 ### Added
 
-- *arch-proto*: added package with protobuf clients [#37](https://github.com/archway-network/arch3.js/pull/37)
+- *arch3-proto*: added package with protobuf clients [#37](https://github.com/archway-network/arch3.js/pull/37)
  
 ### Changed